Voice assistant/requirements.txt
# =============================================================================
# BUDDY Voice Assistant - Python Dependencies
# =============================================================================
# Install with: pip install -r requirements.txt
# For Raspberry Pi 4, some packages may need system dependencies first
# =============================================================================

# -----------------------------------------------------------------------------
# Core Framework
# -----------------------------------------------------------------------------
python-dotenv>=1.0.0         # Environment variable management

# -----------------------------------------------------------------------------
# Audio I/O
# -----------------------------------------------------------------------------
PyAudio>=0.2.13              # Audio recording/playback
sounddevice>=0.4.6           # Alternative audio interface
numpy>=1.24.0                # Audio processing
scipy>=1.10.0                # Signal processing

# -----------------------------------------------------------------------------
# Speech Recognition (faster-whisper)
# -----------------------------------------------------------------------------
faster-whisper>=0.10.0       # Optimized Whisper for CPU
# Note: This automatically installs ctranslate2

# -----------------------------------------------------------------------------
# Wake Word Detection
# -----------------------------------------------------------------------------
openwakeword>=0.5.0          # Open source wake word detection
# Alternative: pvporcupine (requires API key)

# -----------------------------------------------------------------------------
# Voice Activity Detection
# -----------------------------------------------------------------------------
webrtcvad>=2.0.10            # Voice activity detection

# -----------------------------------------------------------------------------
# Text-to-Speech
# -----------------------------------------------------------------------------
# Piper TTS is installed separately as a binary
# See README.md for installation instructions

# Fallback TTS option
pyttsx3>=2.90                # Offline TTS fallback

# -----------------------------------------------------------------------------
# Gemini API
# -----------------------------------------------------------------------------
google-generativeai>=0.3.0   # Google Gemini API client
aiohttp>=3.9.0               # Async HTTP for API calls

# -----------------------------------------------------------------------------
# Utilities  
# -----------------------------------------------------------------------------
asyncio-throttle>=1.0.2      # Rate limiting for API calls
rich>=13.0.0                 # Beautiful terminal output
pydantic>=2.0.0              # Configuration validation
pydantic-settings>=2.0.0     # Settings management

# -----------------------------------------------------------------------------
# Optional: Development & Testing
# -----------------------------------------------------------------------------
# pytest>=7.0.0              # Testing framework
# pytest-asyncio>=0.21.0     # Async test support
# black>=23.0.0              # Code formatter
# ruff>=0.1.0                # Fast linter
